requirements.txt
streamlit>=1.20
scikit-learn>=1.2
pandas>=1.5
numpy>=1.24
joblib>=1.2
scipy>=1.10
pdfminer.six>=20221105
reportlab>=3.6